requirements.txt
beautifulsoup4==4.12.3
boto3==1.34.105
botocore==1.34.105
bs4==0.0.2
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.1
six==1.16.0
soupsieve==2.5
urllib3==2.2.1
